Add files via upload
The next iteration of the dummy tax calculator is similar to the first iteration but calculates federal tax and AZ tax with marginal tax amounts. The output table is very similar to the first, but shows AZ tax and Federal Tax amounts instead of rates. This is more accurate than the previous project 4 code.