Add files via upload
Project 6 is the most recent iteration of the dummy tax calculator, this project demonstrates taking a text file as input and creating multiple text files as an output using a while loop. This project uses additional header files as well as a the Arizona_Tax_Utility.h and Arizona_Tax_Utility.c which I personally created which is used to calculate tax amounts.